requirements.txt
numpy
pandas
tensorflow
keras
pillow
matplotlib
nltk
requests
Flask
